+++ circe/src/test/scala/morphling/circe/CirceSpec.scala
@@ -33,9 +33,35 @@ class CirceSpec extends AnyFunSuite with Matchers with EitherValues with Validat
     )
   }
 
+  test("A value should serialise to JSON [deannotated]") {
+    implicit val encoder = Person.deannotatedSchema.encoder
+
+    person.asJson shouldBe Json.obj(
+      "updateCounter" := 42,
+      "roles" := Seq(
+        Json.obj(
+          "administrator" -> Json.obj(
+            "subordinateCount" := 0,
+            "department" := "windmill-tilting"
+          )
+        )
+      ),
+      "birthDate" := 20147028000L,
+      "name" := "Kris Nuttycombe"
+    )
+  }
+
   test("A value should be deserialised from JSON"){
     implicit val encoder = Person.schema.encoder
     val decoder = Person.schema.decoder
+
+    decoder.decodeJson(person.asJson).value shouldBe person.copy(stamp = 101)
+    decoder.decodeAccumulating(person.asJson.hcursor).value shouldBe person.copy(stamp = 101)
+  }
+
+  test("A value should be deserialised from JSON [deannotated]"){
+    implicit val encoder = Person.deannotatedSchema.encoder
+    val decoder = Person.deannotatedSchema.decoder
 
     decoder.decodeJson(person.asJson).value shouldBe person.copy(stamp = 101)
     decoder.decodeAccumulating(person.asJson.hcursor).value shouldBe person.copy(stamp = 101)
@@ -49,15 +75,35 @@ class CirceSpec extends AnyFunSuite with Matchers with EitherValues with Validat
     decoder.decodeAccumulating(person.asJson.mapObject(_.filterKeys(_ != "updateCounter")).hcursor).value shouldBe person.copy(updateCounter = 0, stamp = 101)
   }
 
+  test("A default value should be applied during deserialization [deannotated]") {
+    implicit val encoder = Person.deannotatedSchema.encoder
+    val decoder = Person.deannotatedSchema.decoder
+
+    decoder.decodeJson(person.asJson.mapObject(_.filterKeys(_ != "updateCounter"))).value shouldBe person.copy(updateCounter = 0, stamp = 101)
+    decoder.decodeAccumulating(person.asJson.mapObject(_.filterKeys(_ != "updateCounter")).hcursor).value shouldBe person.copy(updateCounter = 0, stamp = 101)
+  }
+
   test("Serialization should round-trip values produced by a generator"){
     implicit val arbPerson : Arbitrary[Person] = Arbitrary(Person.schema.gen)
     implicit val encoder = Person.schema.encoder
     val decoder = Person.schema.decoder
     check {
-      (p: Person) => decoder.decodeJson(p.asJson).toOption == Some(p)
+      (p: Person) => decoder.decodeJson(p.asJson).toOption.contains(p)
     }
     check {
-      (p: Person) => decoder.decodeAccumulating(p.asJson.hcursor).toOption == Some(p)
+      (p: Person) => decoder.decodeAccumulating(p.asJson.hcursor).toOption.contains(p)
+    }
+  }
+
+  test("Serialization should round-trip values produced by a generator [deannotated]"){
+    implicit val arbPerson : Arbitrary[Person] = Arbitrary(Person.deannotatedSchema.gen)
+    implicit val encoder = Person.deannotatedSchema.encoder
+    val decoder = Person.deannotatedSchema.decoder
+    check {
+      (p: Person) => decoder.decodeJson(p.asJson).toOption.contains(p)
+    }
+    check {
+      (p: Person) => decoder.decodeAccumulating(p.asJson.hcursor).toOption.contains(p)
     }
   }
 
@@ -90,7 +136,7 @@ class CirceSpec extends AnyFunSuite with Matchers with EitherValues with Validat
     implicit val encoder = Person.flatSchema.encoder
     val decoder = Person.flatSchema.decoder
     check {
-      (p: Person) => decoder.decodeJson(p.asJson).toOption == Some(p)
+      (p: Person) => decoder.decodeJson(p.asJson).toOption.contains(p)
     }
   }
 }